requirements.txt
google-generativeai>=0.3.0
chromadb>=0.4.18
langchain>=0.0.267
sentence-transformers>=2.2.2
PyPDF2>=3.0.0
python-docx>=1.0.0
python-dotenv>=1.0.0
pathlib>=1.0.1
typing>=3.7.4
tqdm>=4.66.0
markdown>=3.4.0
weasyprint>=59.0